requirements.txt
flask
pymongo
flask_mail
python-dotenv
